Fix windows being shown on the wrong workspace
The full layout created a situation in which the switch to another
workspace triggered an exposure event on a window in the previous
workspace which moves moves the focus back again.

A switch involves hiding all windows in the previous workspace but given
that the full layout has all windows overlapping eachother hiding the
top most window causes the window below to fire an exposure event.

Hiding the focused window last fixes this edge case